moved func to plotting.py
