+++ analysis/Python_scripts/utils.py
@@ -1,7 +1,5 @@
 import pandas as pd
-import seaborn as sns
 import numpy as np
-from matplotlib import pyplot as plt
 from rdkit import Chem
 from itertools import combinations
 
@@ -82,68 +80,6 @@ def generate_combinations(df, column_name):
             new_rows.append(new_row)
     
     return pd.DataFrame(new_rows).reset_index(drop=True)
-
-def create_plot(df, path):
-    sns.set_style(style='white')
-    plt.figure(figsize=(17, 5))
-
-    # Set the color palette
-    colors = ['yellow', 'deepskyblue']
-    sns.set_palette(sns.color_palette(colors))
-
-    ax = sns.boxplot(x="true_names", y="value", hue="Number", 
-                    data=df, hue_order=['CosineHungarian_0.01_0.0_1.0_matches',np.nan],
-                    medianprops={'color': 'darkgreen', 'linewidth': 4.0},
-                    flierprops={'marker': 'o', 'markersize': 10, 'markerfacecolor': 'none'})  # RUN PLOT   
-    ax2 = ax.twinx()
-
-    sns.boxplot(ax=ax2,x='true_names', y='value', hue='Number',
-                data=df, hue_order=[np.nan, 'CosineHungarian_0.01_0.0_1.0_scores'], 
-                medianprops={'color': 'b', 'linewidth': 4.0}, 
-                flierprops={'marker': 'o', 'markersize': 10, 'markerfacecolor': 'none'})
-
-    ax.legend_.remove()
-    ax.set_ylim([0, 5])  # Set y-axis limits
-    ax.yaxis.set_major_locator(plt.MultipleLocator(1))  # Set major tick marks
-    ax.set_ylabel('Match values')  # Set y-axis label
-    ax.yaxis.label.set_size(20)  # Set font size of y-axis label
-    ax.set_xlabel('Chemical composition', fontsize=20)  # Set x-axis label and font size
-    ax.tick_params(axis='x', labelsize=13)  # Set font size of x-axis tick labels
-    ax.tick_params(axis='y', labelsize=13)  # Set font size of y-axis tick labels
-    ax.yaxis.labelpad = 10
-    ax.xaxis.labelpad = 10
-
-    # Create a count for each x-axis label
-    count_data = df['true_names'].value_counts().reset_index()
-    count_data.columns = ['true_names', 'count']
-    count_data = count_data.sort_values(by=['true_names'])
-    count_data['count'] = count_data['count'] // 2
-
-    # Remove the original x-axis labels
-    ax.set_xticklabels([])
-
-    # Add the count labels to the x-axis
-    ax.set_xticks(np.arange(len(count_data)))
-    ax.set_xticklabels(count_data['true_names'] + ' (' + count_data['count'].astype(str) + ')', rotation=45, ha='right')
-
-    ax2.legend_.remove()
-    ax2.set_ylim([0, 1])  # Set y-axis limits
-    ax2.yaxis.set_major_locator(plt.MultipleLocator(0.2))  # Set major tick marks
-    ax2.set_ylabel('Score values')  # Set y-axis label
-    ax2.yaxis.label.set_size(20)  # Set font size of y-axis label
-    ax2.tick_params(axis='y', labelsize=12)  # Set font size of y-axis tick labels
-    ax2.yaxis.labelpad = 10
-
-    # Change the legend labels
-    handles, labels = ax.get_legend_handles_labels()
-    labels[0] = 'Matches'
-    labels[1] = 'Scores'
-    ax.legend(handles, labels, loc='upper right', fontsize=14)
-
-    plt.savefig(path, dpi=300, bbox_inches='tight')  # SAVE PLOT AS PNG FILE
-    plt.show()
-    plt.clf()
-    plt.close()
 
 def preprocess_data(merged_top5_same):
     # Concatenate the DataFrames in df1_list and add a 'value' column with the value 'matches'.
